refactor: extract enable_desktop_dpi_awareness into dpi_win.{hpp,cpp}
Self-contained 65-line Win32 fallback chain
(SetProcessDpiAwarenessContext -> SetProcessDpiAwareness -> SetProcessDPIAware)
with zero state shared with the rest of desktop/main.cpp. Mirrors the
existing tray_icon_win / notifications_win / folder_picker_win convention
(_WIN32-guarded TU under namespace acecode::desktop).

Pure code reorganization, no behavior change.
